refactor: split different code gen tests
